Alle sharpes geroggelt

+++ lab3/catmarkframework/CatmarkSubdiv/halfedge.h
@@ -14,7 +14,7 @@ public:
   HalfEdge* twin;
   Face* polygon;
   unsigned int index;
-  unsigned short sharpness;
+  float sharpness;
 
   // Inline constructors
